chore: delete dead scripts and dedup the UTC time helpers
- Remove four scripts with no caller in any workflow, script, doc or test.
  retry_rejects_phone.py is superseded by retry_rejects_cached.py, which
  CLAUDE.md already recommends instead; the other three were untouched since
  April and referenced nowhere.
- Remove two committed console dumps that no script produces any more.
  love-dining-verification.txt stays because verify_love_dining.py writes it.
- Collapse two identical iso_now definitions and four hand-rolled lenient ISO
  parses into scripts/timeutil.py. The strict, field-named parsers stay where
  they are: their error messages are asserted by the receipt tests and are the
  point of those helpers.
- parse_utc rejects a naive timestamp rather than assuming UTC. Guessing the
  zone is how a stale record reads as fresh.
- Drop the unused Action.SKIP added earlier today.
