--- data/audit-inactive-restaurants.txt
@@ -1,221 +1,0 @@
-================================================================================
-GLOBAL DINING AUDIT REPORT
-Comparing local data against public country listing pages
-================================================================================
-
-Total in local data: 2413
-Verified in public listings: 0
-Not found in public listings: 2413
-Unverified/unknown: 0
-
-RESTAURANTS NOT IN PUBLIC LISTINGS (candidates for removal):
---------------------------------------------------------------------------------
-
-Australia (162 inactive):
-  - Chairman & Yip (Barton)
-  - Meat & Wine Co Canberra (Canberra)
-  - Onzieme (Kingston)
-  - Pilot. (Ainslie)
-  - RAKU (Canberra)
-  - Terra (Canberra)
-  - Wilma (Canberra)
-  - 6HEAD (The Rocks)
-  - Aria (Sydney)
-  - Armorica (Surry Hills)
-  ... and 152 more
-
-Austria (166 inactive):
-  - Lounge 81 (Bad Gleichenberg)
-  - Grand Café & Restaurant Esplanade (Bad Ischl)
-  - Triad (Bad Schönau)
-  - El Gaucho Im Josefsbad (Baden)
-  - Wein & Co Restaurant & Shops (Bregenz)
-  - Weiss Restaurant (Bregenz)
-  - Gourmet Restaurant Hubert Wallner (Dellach/Maria Wörth)
-  - Zum Verwalter (Dornbirn)
-  - DER TSCHEBULL (Egg am Faaker See)
-  - Die Weinbank (Ehrenhausen an der Weinstraße)
-  ... and 156 more
-
-Canada (333 inactive):
-  - Alloy (Calgary)
-  - Bar Gigi (Calgary)
-  - Bridgette Bar (Calgary)
-  - Calcutta Cricket Club (Calgary)
-  - charbar (Calgary)
-  - Charcut (Calgary)
-  - D.O.P. (Calgary)
-  - Deane House (Calgary)
-  - DOPO (Calgary)
-  - EIGHT (Calgary)
-  ... and 323 more
-
-France (302 inactive):
-  - Mamo Michelangelo (Antibes)
-  - Restaurant Louroc (CS 10029)
-  - African queen (Beaulieu-sur-Mer)
-  - Le Cent 33 (Bordeaux)
-  - Ressources (Bordeaux)
-  - La Guérite (Cannes)
-  - La Môme Cannes (Cannes)
-  - La Môme Plage (Cannes)
-  - Midô Cannes (Cannes)
-  - Palme d'or (Cannes)
-  ... and 292 more
-
-Germany (283 inactive):
-  - ALEX Aachen Markt (Aachen)
-  - BLOCK HOUSE (Ahrensburg)
-  - Bell’s Steakhouse (Aschheim)
-  - BLOCK HOUSE (Augsburg)
-  - 1811 (Berlin)
-  - 893 Ryotei (Berlin)
-  - ALEX Berlin am Alex (Berlin)
-  - ALEX Berlin am Mercedes Platz (Berlin)
-  - BLOCK HOUSE (Berlin)
-  - BLOCK HOUSE (Berlin)
-  ... and 273 more
-
-Hong Kong (69 inactive):
-  - Prohibition Grill house & Cocktail bar (180 Wong Chuk Hang Road)
-  - Lobster Bar and Grill (Supreme Court Road)
-  - Nadaman (Island Shangri-la) (Supreme Court Road)
-  - Nicholini’s (88 Queensway)
-  - Summer Palace (Supreme Court Road)
-  - Alto Bar & Grill (18 Tang Lung Street)
-  - Kanizen (33 Hysan Avenue)
-  - Roganic (33 Hysan Avenue)
-  - Shimo (15 Sharp Street East)
-  - The Drunken Pot (V Point) (18 Tang Lung Street)
-  ... and 59 more
-
-Italy (59 inactive):
-  - La Caravella (Amalfi)
-  - I Portici (Bologna)
-  - Da Vittorio (Brusaporto)
-  - Taverna Estia (Brusciano)
-  - Osteria del Forte (Cagliari)
-  - Reale Niko Romito (Castel Di Sangro)
-  - La Bandiera (Civitella Casanova)
-  - Gucci Osteria Da Massimo Bottura (Florence)
-  - Il Palagio (99)
-  - Ora D'Aria (Florence)
-  ... and 49 more
-
-Mexico (373 inactive):
-  - Becco Al Mare (Playa Guitarron)
-  - Fisher's Acapulco (1926 Local T-2)
-  - Shu Acapulco (Fracc Playa Diamante)
-  - Suntory Costera (Costa Azul)
-  - Zibu (Fracc Glomar)
-  - Harrys Cabo (Tourist Corridor)
-  - Ilios Cabo (Tourist Corridor)
-  - Nicksan Cabo (Cabo San Lucas)
-  - Yaya (Cabo San Lucas)
-  - Argentino Cambalache Cancún (Cancún)
-  ... and 363 more
-
-Monaco (10 inactive):
-  - Abysse Monte-Carlo (1 square Beaumarchais)
-  - Amazónico Monte-Carlo (Monaco)
-  - Beefbar Monaco (Monaco)
-  - Blue Bay Marcel Ravin (Monaco)
-  - Buddha-Bar Monte-Carlo (Monaco)
-  - Le Louis XV - Alain Ducasse à l’Hôtel de Paris (Monaco)
-  - Marlow (Monaco)
-  - Maya Bay (Monaco)
-  - Pavyllon Monte-Carlo (1 square Beaumarchais)
-  - Song Qi (Monaco)
-
-New Zealand (70 inactive):
-  - Aosta (Arrowtown)
-  - Ahi (7/21 Queen Street)
-  - Alma (Auckland)
-  - Amano (Auckland CBD)
-  - Andiamo (Ponsonby)
-  - Baduzzi (North Wharf)
-  - Bar Magda (Auckland)
-  - Botswana Butchery - Auckland (99 Quay Street)
-  - Café Hanoi (Britomart)
-  - Cassia (Auckland CBD)
-  ... and 60 more
-
-Singapore (65 inactive):
-  - Araya (Singapore)
-  - Asian Market Café (Fairmont Singapore)
-  - Basilico (Singapore)
-  - Basque Kitchen by Aitor (The Fullerton Waterboat House)
-  - Bee's Knees (Singapore)
-  - Boon Tong Kee (Balestier Road) (Singapore)
-  - Botanico (Singapore)
-  - Brasserie Astoria (Singapore)
-  - Burnt Ends (Singapore)
-  - Butcher's Block (Raffles Arcade)
-  ... and 55 more
-
-Spain (51 inactive):
-  - Bastian Beach Club (Ciutat Vella)
-  - Carpe Diem Lounge Club (32)
-  - Cecconi’s Barcelona (4)
-  - Cocina Hermanos Torres (Les Corts)
-  - Lomo Alto (Barcelona)
-  - Lomo Bajo (ground floor)
-  - Rest Botafumeiro (81)
-  - Chezz Gerdi Formentera (Formentera)
-  - Restaurante Juan Y Andrea (Formentera)
-  - Cala Bassa Beach Club (Sant Josep De Sa Talaia)
-  ... and 41 more
-
-Taiwan (104 inactive):
-  - Go eat Tapas Dining BAR (East Dist.)
-  - Ichiyaki (East Dist.)
-  - G.D Restaurant & Bar (Zhubei City)
-  - Wang Steak - Zhubei Guangming Branch (Zhubei City)
-  - Chun Xi Da Bian Lu Kaohsiung Mingcheng Branch (Zuoying Dist)
-  - Chun Xi Da Bian Lu Kaohsiung Wenbin Branch (Fengshan Dist.)
-  - Crab's House (Xinxing Dist.)
-  - HAILI (Qianjin District)
-  - Liberté (Qianzhen Dist.)
-  - Piccolo Polpo Bistro (Kaohsiung) (Gushan Dist.)
-  ... and 94 more
-
-Thailand (80 inactive):
-  - 1823 Tea Lounge by Ronnefeldt (Pathumwan)
-  - 80/20 (Bangrak)
-  - Acqua (Pathumwan)
-  - Aesops Greek Restaurant and Rooftop (Khlongtoei)
-  - Aksorn (Bangrak)
-  - Allday Canteen (Bangkok)
-  - Baan Dusit Thani (Bangkok)
-  - Blue by Alain Ducasse (Khlongsan)
-  - Blue Elephant Restaurant (Bangkok)
-  - Brasserie 9 (Bangrak)
-  ... and 70 more
-
-United Kingdom (208 inactive):
-  - The Ivy Bath Brasserie (Bath)
-  - The Ivy Belfast Brasserie (56 Donegall Pl)
-  - 24 Stories (Birmingham)
-  - 670 grams (Deritend)
-  - Dishoom Birmingham (Birmingham)
-  - Opheem (Birmingham)
-  - Simpsons (Edgbaston)
-  - Tattu Birmingham (Birmingham)
-  - The Ivy Temple Row (Birmingham)
-  - Trillium (Birmingham)
-  ... and 198 more
-
-United States (78 inactive):
-  - Atrium (Atlanta)
-  - Aziza (Atlanta)
-  - Bar Vegan (Atlanta)
-  - Twisted Soul Cookhouse & Pours (Atlanta)
-  - Orfano (Boston)
-  - Rochambeau (Boston)
-  - Scampo (Boston)
-  - Sonsie (Boston)
-  - Franks Wine Bar (Brooklyn)
-  - Maison Yaki (Brooklyn)
-  ... and 68 more
-
-Total candidate removals: 2413
